Many (big) changes here
1. What you're passing to create read, create relay, etc. has been changed to
ReadConfig and RelayConfig. Responses for many read and relay operations
also return configs (instead of "relays", "reads" or whatever else.
2. Many "internal" fields have been prefixed with a "_" - this will cause them
to become exposed in compiled protobuf as XSomething. This is good - it
signifies that the field is intended for _private_ usage (similar to python).
3. ReadRecord now contains a private _decoded field which will be used by
plumber to store.. decoded values :)
4. WriteRecord now contains a private _encoded field which will be used by
plumber to store.. encoded values :)
5. Some indexes have changed - things will be broken. Sorry :/
6. ReadRecord now contains a "metadata" which is a map<string, bytes>

ALL of this is because it has become apparent to me that plumber should have a
SINGLE source of truth for read/write messages. Currently (as I type this),
plumber has `*types.ReadMessage` and `*types.WriteMessage`. The plan is that
with this change - we'll be able to get rid of those types and ONLY use the
proto types. It will simplify the codebase and cause less confusion.

For example - we'll no longer need to have converter/generator methods that
take a protobuf WriteRequest and try to create a *types.WriteMessage out of it.

Another bonus: we now have exactly ONE place to define what properties/fields
a particular backend record has - in the proto schema (and NOT in plumber
itself).

:fingers_crossed: