more big changes
It is now possible to pass a connection_config as part of a read, write or relay
request. While in server mode, we will be passing a connection_id, in CLI, we
need to be able to pass an ephemeral connection config.

In other words, connection_id takes precedence - if not set, we will look
for a connection config.